1. 댓글 코멘트, 코드 분리 백엔드 수정 /

+++ teamKMS/src/main/java/com/devworker/kms/repo/common/CommentRepo.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.devworker.kms.entity.common.BoardDao;
 import com.devworker.kms.entity.common.CommentDao;
 
-public interface CommentRepo extends JpaRepository<CommentDao, Integer> {
+public interface CommentRepo extends JpaRepository<CommentDao, Long> {
 
 	List<CommentDao> findByBoardId(BoardDao boardObj);
 }
